requirements/development.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.3
coreapi==2.3.1
coreschema==0.0.4
coverage==4.4.1
Django==1.11.1
django-rest-swagger==2.1.2
djangorestframework==3.6.3
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
mock==2.0.0
openapi-codec==1.3.1
packaging==16.8
pbr==3.0.1
py==1.4.33
PyMySQL==0.7.11
pyparsing==2.2.0
pytest==3.1.1
pytest-django==3.1.2
python-logstash==0.4.6
pytz==2017.2
requests==2.17.3
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1